move handle_groups to service
